Move some types from ScheduleFinder component
