+++ apps/site/assets/ts/schedule/components/ScheduleFinder.tsx
@@ -5,7 +5,9 @@ import {
   SimpleStopMap,
   RoutePatternsByDirection,
   ServiceInSelector,
-  ScheduleNote as ScheduleNoteType
+  ScheduleNote as ScheduleNoteType,
+  SelectedDirection,
+  SelectedOrigin
 } from "./__schedule";
 import { handleReactEnterKeyPress } from "../../helpers/keyboard-events";
 import icon from "../../../static/images/icon-schedule-finder.svg";
@@ -24,16 +26,6 @@ interface Props {
   routePatternsByDirection: RoutePatternsByDirection;
   today: string;
   scheduleNote: ScheduleNoteType | null;
-}
-
-export type SelectedDirection = 0 | 1 | null;
-export type SelectedOrigin = string | null;
-
-export interface UserInput {
-  route: string;
-  origin: string;
-  date: string;
-  direction: SelectedDirection;
 }
 
 interface State {
